backend/requirements.txt
django
djangorestframework
requests
beautifulsoup4
python-dotenv
openai
pymongo
